+++ packages/instant-apps-components/src/components.d.ts
@@ -888,8 +888,10 @@ export namespace Components {
         "titleText": string;
     }
     interface InstantAppsTimeFilter {
+        "autoPlay": boolean;
         "filterMode": FilterMode;
         "timeInfoConfigItems": ITimeInfoConfigItem[];
+        "timeSliderConfig": __esri.widgetsTimeSliderProperties;
         "view": __esri.MapView | __esri.SceneView;
     }
 }
@@ -2271,8 +2273,10 @@ declare namespace LocalJSX {
         "titleText"?: string;
     }
     interface InstantAppsTimeFilter {
+        "autoPlay"?: boolean;
         "filterMode"?: FilterMode;
         "timeInfoConfigItems"?: ITimeInfoConfigItem[];
+        "timeSliderConfig"?: __esri.widgetsTimeSliderProperties;
         "view"?: __esri.MapView | __esri.SceneView;
     }
     interface IntrinsicElements {