+++ packages/instant-apps-components/CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.0.0-beta.246
+
+### instant-apps-time-filter
+
+- fix: time slider prop changes
+
+## v1.0.0-beta.245
+
+### instant-apps-time-filter
+
+- feat: Add support for auto play and time slider config
+
+## v1.0.0-beta.244
+
+### General
+
+- chore: update calcite to 2.11.1
+
 ## v1.0.0-beta.243
 
 ### instant-apps-scoreboard